Attempts at making the docs more clear with secret volume
The secret includes keys/value pair and a secret, when mounted as a
volume, will generate a file for each of the pair where the name of the
file is the key and the content of the file, the value.

This hopefully makes the doc clear on how to configured credentials.